python-service/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# AI Provider SDKs
anthropic==0.39.0
openai==1.54.0

# ML and Data Science
prophet==1.1.6
pandas==2.2.3
numpy==2.1.3

# Database
supabase==2.9.1

# Scheduling
apscheduler==3.10.4

# Environment and Configuration
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.0

# Security
cryptography==43.0.3

# HTTP and CORS
httpx==0.27.2
